Add tests for removing project and user
