+++ frontend/test/integration/student/studentList.test.tsx
@@ -1,15 +1,15 @@
-import { StudentList } from "../../src/components/student/studentList";
+import { StudentList } from "../../../src/components/student/studentList";
 import "@testing-library/jest-dom";
 import { render, RenderResult, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import mockAxios from "jest-mock-axios";
 import { AxiosResponse } from "axios";
-import { enableUseEditionComponentWrapper, getAxiosCallWithEdition, makeCacheFree } from "./Provide";
-import { getBaseActiveEdition, getBaseOkResponse, getBasePage, getBaseStudent } from "./TestEntityProvider";
-import apiPaths from "../../src/properties/apiPaths";
-import { studentCollectionName } from "../../src/api/entities/StudentEntity";
+import { enableUseEditionComponentWrapper, getAxiosCallWithEdition, makeCacheFree } from "../Provide";
+import { getBaseActiveEdition, getBaseOkResponse, getBasePage, getBaseStudent } from "../TestEntityProvider";
+import apiPaths from "../../../src/properties/apiPaths";
+import { studentCollectionName } from "../../../src/api/entities/StudentEntity";
 import mockRouter from "next-router-mock";
-import { getQueryUrlFromParams } from "../../src/api/calls/baseCalls";
+import { getQueryUrlFromParams } from "../../../src/api/calls/baseCalls";
 
 jest.mock("next/router", () => require("next-router-mock"));
